fix: resolve 'Invalid column number (column -1 requested)' error
When processing eslint-disable-line or eslint-disable-next-line comment directives
within CSS, the code was setting column value to -1. This caused an error when
ESLint's getIndexFromLoc method was called with this invalid column number.

Changed column value from -1 to 0 to indicate the start of the line, which is
a valid column position.

Fixes parsing errors that occur at <style scoped> tags in Vue SFC files.